# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TasksApp API",
        "description": "Documentación de la API de TasksApp",
        "contact": {
            "name": "Nombre del contacto",
            "url": "http://example.com/contact",
            "email": "contact@example.com"
        },
        "license": {
            "name": "Licencia",
            "url": "http://example.com/license"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/auth/login": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Autenticar usuario y generar token",
                "operationId": "776b33040823090b71e359d40afc8f26",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "12345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login exitoso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "1|qwertyuiopasdfghjklzxcvbnm123456"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Credenciales incorrectas"
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Cerrar sesión y revocar token",
                "operationId": "3f4b1ed9ff274eaa8b6c87ae38a4ba50",
                "responses": {
                    "200": {
                        "description": "Usuario desautenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "email": {
                                            "type": "string",
                                            "example": "Usuario desautenticado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/auth/register": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Registrar un nuevo usuario",
                "operationId": "ed42f392010c301892e640067141676e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Juan Pérez"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "juan@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "12345678"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "12345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuario registrado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Juan Pérez"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "juan@example.com"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2025-08-11T08:00:00.000000Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "example": "2025-08-11T08:00:00.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "1|asdfghjkl1234567890qwertyuiop"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Datos inválidos"
                    }
                }
            }
        },
        "/proyects": {
            "get": {
                "tags": [
                    "Proyectos"
                ],
                "summary": "Listar proyectos",
                "operationId": "32f43d2bfc8e0ea3db4d68d653e8f5ca",
                "responses": {
                    "200": {
                        "description": "Lista de proyectos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "model": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Proyect"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "204": {
                        "description": "No hay proyectos"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Proyectos"
                ],
                "summary": "Crear un nuevo proyecto",
                "operationId": "a5b3ab8a6592db9209afba7676f16c8f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Sistema de inventario"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Proyecto para control de stock",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Proyecto creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Proyecto creado exitosamente"
                                        },
                                        "model": {
                                            "$ref": "#/components/schemas/Proyect"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Datos inválidos"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/proyects/{id}": {
            "get": {
                "tags": [
                    "Proyectos"
                ],
                "summary": "Obtener un proyecto",
                "operationId": "386a091339b6f4dede758999ebb7cffc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del proyecto",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Proyecto encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "model": {
                                            "$ref": "#/components/schemas/Proyect"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Proyecto no encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Proyectos"
                ],
                "summary": "Actualizar un proyecto",
                "operationId": "26107a37a3a8c1c73a712f83a98f5e5e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del proyecto",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Sistema de ventas"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Proyecto para registrar ventas",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Proyecto actualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "model": {
                                            "$ref": "#/components/schemas/Proyect"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Proyecto no encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Proyectos"
                ],
                "summary": "Eliminar un proyecto",
                "operationId": "aad986c5cf64fd25f7a397514e25e7c0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del proyecto",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Proyecto eliminado correctamente"
                    },
                    "404": {
                        "description": "Proyecto no encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Proyect": {
                "title": "Proyect",
                "required": [
                    "name"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Sistema de gestión"
                    },
                    "description": {
                        "type": "string",
                        "example": "Proyecto para administrar tareas internas",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-08-11T09:00:00.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-08-11T09:00:00.000000Z"
                    },
                    "deleted_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": null,
                        "nullable": true
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Autenticación",
            "description": "Autenticación"
        },
        {
            "name": "Proyectos",
            "description": "Proyectos"
        }
    ]
}